Refactor: Extract shared utils and simplify search params handling
